Day01/requirements.txt
flake8==7.0.0
pandas==2.2.2
psycopg2-binary==2.9.9